simple_chatbot/requirements.txt
streamlit
pandas
requests
gspread
google-auth
google-generativeai
python-dotenv